add Edit::ReplaceAll variant
this variant is implemented by operating directly on the existing buffer, which is much more optimal than recreating the buffer from scratch, or, for example, combining SelectAll + Paste on an exisitng buffer. set_rich_text is  used because, at time of writing, it is optimised for reusing any existing allocations that it can.